backend/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.14.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.3.0
auth0-python==4.8.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
build==1.2.2.post1
cachetools==6.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.2.1
cloudpickle==3.1.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
configparser==7.2.0
contourpy==1.3.1
courlan==1.3.2
crewai==0.102.0
cryptography==44.0.1
cssselect==1.3.0
cycler==0.12.1
databricks-sdk==0.39.0
dataclasses-json==0.6.7
dateparser==1.2.2
debugpy==1.8.8
decorator==5.1.1
Deprecated==1.2.15
distlib==0.3.9
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
durationpy==0.9
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.1.0
fastapi==0.116.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.55.1
frozenlist==1.5.0
fsspec==2025.2.0
fst-pso==1.8.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
google-pasta==0.2.0
google-search-results==2.4.2
googleapis-common-protos==1.68.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
groq==0.31.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h5py==3.12.1
htmldate==1.9.3
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
instructor==1.7.2
ipykernel==6.29.5
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.8.2
joblib==1.5.1
json5==0.10.0
json_repair==0.39.0
jsonpatch==1.33
jsonpickle==4.0.2
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jusText==3.0.2
keras==3.7.0
kiwisolver==1.4.7
kubernetes==32.0.1
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-groq==0.3.7
langchain-huggingface==0.3.1
langchain-ollama==0.3.7
langchain-openai==0.3.32
langchain-pinecone==0.2.12
langchain-pymupdf4llm==0.4.1
langchain-text-splitters==0.3.11
langdetect==1.0.9
langsmith==0.4.27
libclang==18.1.1
litellm==1.60.2
load-dotenv==0.1.0
lxml==5.4.0
lxml_html_clean==0.4.2
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
miniful==0.0.6
ml-dtypes==0.4.1
mlflow==2.19.0
mlflow-skinny==2.19.0
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy_extensions==1.1.0
namex==0.0.8
narwhals==1.48.0
nest-asyncio==1.6.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.6
oauthlib==3.2.2
ollama==0.5.3
onnxruntime==1.20.1
openai==1.107.0
openpyxl==3.1.5
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.3.1
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.3.0
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
plotly==6.2.0
posthog==3.15.0
prompt_toolkit==3.0.48
propcache==0.3.0
proto-plus==1.26.0
protobuf==6.31.1
psutil==6.1.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pyFUME==0.3.4
Pygments==2.18.0
PyJWT==2.10.1
PyMuPDF==1.26.4
pymupdf4llm==0.0.27
pyparsing==3.2.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyvis==0.3.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.26.0
rsa==4.9
safetensors==0.6.2
schedule==1.2.2
scikit-learn==1.7.1
scipy==1.15.3
sentence-transformers==5.1.0
sgmllib3k==1.0.0
shellingham==1.5.4
simpful==2.12.0
simsimd==6.5.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.47.2
streamlit==1.47.0
sympy==1.13.3
tenacity==9.1.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tinysegmenter==0.3
tld==0.13.1
tldextract==5.3.0
tokenizers==0.22.0
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
torch==2.8.0
tornado==6.5.1
tqdm==4.67.1
trafilatura==2.0.0
traitlets==5.14.3
transformers==4.56.1
typer==0.15.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.5.0
uv==0.6.2
uvicorn==0.35.0
virtualenv==20.27.1
waitress==3.0.2
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0
Werkzeug==3.1.3
wrapt==1.17.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
